Add translate dialog for users with other languages
