+++ frontend/src/components/Egg/Egg.module.scss
@@ -5,10 +5,14 @@
   outline: none;
   width: 100%;
   height: 100%;
-  display: flex;
+  display: none;
   align-items: center;
   justify-content: center;
   overflow: visible;
+
+  &[open] {
+    display: flex;
+  }
 
   &::backdrop {
     background: rgba(0,0,0,.6);
